[Discover] Clean up types
