+++ src/plugins/unified_histogram/public/chart/point_series.test.ts
@@ -8,7 +8,7 @@
 
 import { buildPointSeriesData } from './point_series';
 import moment from 'moment';
-import { Unit } from '@kbn/datemath';
+import type { Unit } from '@kbn/datemath';
 
 describe('buildPointSeriesData', () => {
   test('with valid data', () => {
