+++ src/plugins/unified_histogram/public/chart/use_chart_panels.ts
@@ -5,6 +5,7 @@
  * in compliance with, at your election, the Elastic License 2.0 or the Server
  * Side Public License, v 1.
  */
+
 import { i18n } from '@kbn/i18n';
 import type {
   EuiContextMenuPanelItemDescriptor,
