+++ src/plugins/unified_histogram/public/panels/panels_fixed.tsx
@@ -8,7 +8,8 @@
 
 import { EuiFlexGroup, EuiFlexItem } from '@elastic/eui';
 import { css } from '@emotion/react';
-import React, { ReactElement } from 'react';
+import type { ReactElement } from 'react';
+import React from 'react';
 
 export const PanelsFixed = ({
   className,
