+++ src/plugins/unified_histogram/public/plugin.ts
@@ -6,7 +6,7 @@
  * Side Public License, v 1.
  */
 
-import { Plugin } from '@kbn/core/public';
+import type { Plugin } from '@kbn/core/public';
 
 export class UnifiedHistogramPublicPlugin implements Plugin<{}, {}, object, {}> {
   public setup() {
