+++ src/plugins/unified_histogram/public/types.ts
@@ -6,10 +6,74 @@
  * Side Public License, v 1.
  */
 
-import { Theme } from '@kbn/charts-plugin/public/plugin';
-import { IUiSettingsClient } from '@kbn/core/public';
-import { DataPublicPluginStart } from '@kbn/data-plugin/public';
-import { FieldFormatsStart } from '@kbn/field-formats-plugin/public';
+import type { Theme } from '@kbn/charts-plugin/public/plugin';
+import type { IUiSettingsClient } from '@kbn/core/public';
+import type { DataPublicPluginStart } from '@kbn/data-plugin/public';
+import type { FieldFormatsStart } from '@kbn/field-formats-plugin/public';
+import type { Duration, Moment } from 'moment';
+import type { Unit } from '@kbn/datemath';
+import type { SerializedFieldFormat } from '@kbn/field-formats-plugin/common';
+
+interface Column {
+  id: string;
+  name: string;
+}
+
+interface Row {
+  [key: string]: number | 'NaN';
+}
+
+interface HistogramParamsBounds {
+  min: Moment;
+  max: Moment;
+}
+
+interface HistogramParams {
+  date: true;
+  interval: Duration;
+  intervalESValue: number;
+  intervalESUnit: Unit;
+  format: string;
+  bounds: HistogramParamsBounds;
+}
+
+interface Dimension {
+  accessor: 0 | 1;
+  format: SerializedFieldFormat<{ pattern: string }>;
+  label: string;
+}
+
+interface Ordered {
+  date: true;
+  interval: Duration;
+  intervalESUnit: string;
+  intervalESValue: number;
+  min: Moment;
+  max: Moment;
+}
+
+export interface Table {
+  columns: Column[];
+  rows: Row[];
+}
+
+export interface Dimensions {
+  x: Dimension & { params: HistogramParams };
+  y: Dimension;
+}
+
+export interface ChartData {
+  values: Array<{
+    x: number;
+    y: number;
+  }>;
+  xAxisOrderedValues: number[];
+  xAxisFormat: Dimension['format'];
+  yAxisFormat: Dimension['format'];
+  xAxisLabel: Column['name'];
+  yAxisLabel?: Column['name'];
+  ordered: Ordered;
+}
 
 export interface UnifiedHistogramServices {
   data: DataPublicPluginStart;
